# docs/swagger.yaml
definitions:
...
  controllers.MessageResponse:
    properties:
      message:
        type: string
    type: object
  photo.NewPhoto:
    properties:
      caption:
        example: caption
        type: string
      photo_url:
        example: www.photo.com
        type: string
      title:
        example: title
        type: string
      user_id:
        type: integer
    required:
    - photo_url
    - title
    type: object
  photo.PaginationResultPhoto:
    properties:
      current:
        type: integer
      data:
        items:
          $ref: '#/definitions/photo.Photo'
        type: array
      limit:
        type: integer
      nextCursor:
        type: integer
      numPages:
        type: integer
      prevCursor:
        type: integer
      total:
        type: integer
    type: object
  photo.Photo:
    properties:
      caption:
        example: caption
        type: string
      created_at:
        example: "2021-02-24 20:19:39"
        type: string
      deleted_at:
        example: "null"
        type: string
      id:
        example: 1099
        type: integer
      photo_url:
        example: www.photo.com
        type: string
      title:
        example: title
        type: string
      updated_at:
        example: "2021-02-24 20:19:39"
        type: string
      user_id:
        type: integer
    type: object
...
  sosmed.NewSocialMedia:
    properties:
      name:
        example: name
        type: string
      social_media_url:
        example: www.sosmed.com
        type: string
      user_id:
        type: integer
    required:
    - name
    - social_media_url
    type: object
  sosmed.PaginationResultSocialMedia:
    properties:
      current:
        type: integer
      data:
        items:
          $ref: '#/definitions/sosmed.SocialMedia'
        type: array
      limit:
        type: integer
      nextCursor:
        type: integer
      numPages:
        type: integer
      prevCursor:
        type: integer
      total:
        type: integer
    type: object
  sosmed.SocialMedia:
    properties:
      created_at:
        example: "2021-02-24 20:19:39"
        type: string
      deleted_at:
        example: "2021-02-24 20:19:39"
        type: string
      id:
        example: 1099
        type: integer
      name:
        example: caption
        type: string
      social_media_url:
        example: www.sosmed.com
        type: string
      updated_at:
        example: "2021-02-24 20:19:39"
        type: string
      user_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /photo:
    get:
      description: Get all Photos on the system
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/photo.PaginationResultPhoto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all Photos
      tags:
      - photo
    post:
      consumes:
      - application/json
      description: Create new photo on the system
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/photo.NewPhoto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/photo.Photo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Create New PhotoName
      tags:
      - photo
  /photo/{photo_id}:
    get:
      description: Get Photos by ID on the system
      parameters:
      - description: id of photo
        in: path
        name: photo_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get photos by ID
      tags:
      - photo
  /sosmed:
    get:
      description: Get all SocialMedia on the system
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sosmed.PaginationResultSocialMedia'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all SocialMedia
      tags:
      - sosmed
    post:
      consumes:
      - application/json
      description: Create new sosmed on the system
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/sosmed.NewSocialMedia'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sosmed.SocialMedia'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Create New SocialMediaName
      tags:
      - sosmed
  /sosmed/{sosmed_id}:
    get:
      description: Get SocialMedia by ID on the system
      parameters:
      - description: id of sosmed
        in: path
        name: sosmed_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get sosmeds by ID
      tags:
      - sosmed
...
swagger: "2.0"